REST api permissions + bug fixes

+++ src/app/api/class/route.tsx
@@ -1,4 +1,5 @@
 import { createNewClass, deleteClass, getClassesByTeacherId, updateClass } from '@/actions/classes';
+import { checkForUnauthorizedTeacher } from '@/app/lib/utils';
 import { newClassSchema, updateClassSchema } from '@/app/lib/zod';
 import { NextResponse } from 'next/server';
 import { ZodError } from 'zod';
@@ -8,7 +9,7 @@ export async function POST(req: Request) {
         const body = await req.json();
 
         const { class_grade, class_section, female_number, male_number, details, teacher_id } = await newClassSchema.parseAsync(body);
-
+        checkForUnauthorizedTeacher(teacher_id);
         const result = await createNewClass(class_grade, class_section, female_number, male_number, details, teacher_id);
 
         if (result) {
@@ -30,6 +31,8 @@ export async function GET(req: Request) {
     const { searchParams } = new URL(req.url);
     const teacherId = searchParams.get('uid');
 
+    checkForUnauthorizedTeacher(teacherId);
+
     if (!teacherId) {
         return NextResponse.json({ error: 'Missing teacher_id parameter' }, { status: 400 });
     }
@@ -41,9 +44,10 @@ export async function GET(req: Request) {
 export async function PUT(req: Request) {
     try {
         const body = await req.json();
+        checkForUnauthorizedTeacher(body.teacher_id);
         const { female_number, male_number, details } = await updateClassSchema.parseAsync(body);
 
-        const result = await updateClass(body.class_id, male_number, female_number, details)
+        await updateClass(body.class_id, male_number, female_number, details)
         return NextResponse.json({ message: 'Ok' }, { status: 200 });
     } catch (error: any) {
         return NextResponse.json({ message: error }, { status: 400 });
@@ -53,6 +57,7 @@ export async function PUT(req: Request) {
 export async function DELETE(req: Request) {
     try {
         const body = await req.json();
+        checkForUnauthorizedTeacher(body.teacher_id);
         const res = await deleteClass(body.class_id);
         if(res) {
             return NextResponse.json({ data: 'Ok' }, { status: 200 });
